added file save for night directory with user-selected base file name
The user can choose which base filename to use, and also include object name and type, all saved in the fits header.  There is also a counter that keeps track of all the files you've created, and appends that file number to the end of each base name.
Added FITS keywords COMBO and NCOMBO, which say whether an image is a combination of multiple, and how many were combined.
New files are saved in directory for the date of observation: SISI_images/SAMOS_yyyymmdd