fixed a lot of file naming
added some indicator lights to the GUI for when motors are moving (not working yet)

fixed some things with file corrections with "subtract for display".

Added exposure time to dark files.  When looking for darks for subtraction, first check if the exposure time in the superdark file is the same.  If not, get the superdark image that has the closest exposure time.