End of the week
Hadamard masks now in decent shape, added a field to operate them in the main menu and the possibility of defining the center coordinates in the RADEC in the HTS generation panel. Also the RADEC in the display are now presented ALSO in decimal degrees, this is what is used by the HTS centroids.
Still fragile if entries are wrong.