fixes to the HDRs
In WriteFITSHead.py I split the function in two pieces, one for the astrometry and one for the rest; this because incomplete astrometry added to the non science image files still shows RADEC on the image display, which is annoying.
Then I added the RADEC display back to the  Skymapper image BEFORE the twirl solution. Not really needed, but right and good experiece for the other catalogs, to be done