Massimo first check
Comments were already shared on Slack, but I am attaching here again the two files.

--------------
ConfigPage

1) SAMOS Configuration: We have four choices for Store Files. Are they all necessary? The best choice seems to be "Working Directory", but probably we can label it more clearly (e.g. "Default Data Directory") and then give an option for a custom directory.
Note also that we create the folder "SISI Images", but in the near future we hope to be able to store also the SAMI Images, so a more general name can be at the horizon, e.g. "SAMOS Images" divided in "SISI" and "SAMI"

2) Servers
I think that the "Outside (with VPN)" is not needed anymore. But we never know... so one could just comment out the button but leave the code so in the case we can reactivate the mode without too much effort.

3) Flip images on open
It has given us so many problems with the guide stars that I would remove it. Let's leave the images the way they arrive (i.e. flipped in RA) and if one wants to see them with the right orientation, it can always load them in ds9 and do it externally.

4) IP Addresses
According to my notes, the SOAR TCS IP address is 139.220.15.2:40050
while SAMI has 139.229.15.163
We can "hardcode" these two addresses in the list.
=> DONE

5) If if check "SIMULATED", the DMD and SOAR telescope circles turn green. Then, checking back "INSIDE", they stay green. I think they should go black. A secondary question is why only these two circles react to the SIMULATED status.

6) The "Initialize Components" initialize everything. I assume that if a component is not connected we get some diagnostic but the program keeps running; otherwise we need to have check buttons for the peripherals we want to talk to, so we can unselect those that may give us problems if unconnected.

7) Having selected "INSIDE" to see what happens, I decided to exit the program.
Then I restarted, and the code had now "INSIDE" preselected. However, there are no motors connected and a socket error was issued. It makes a check every 180s and runs for ~55minutes before the simulator shuts down. Need to be fixed.

8) Logbook is created in a rather arcane subdirectory, e.g.
./src/samos/data/tmp/SISI_images
I wonder if the logbook can go in the same directory of the data, or somewhere higher up where one can more easily download it at the end of the night.

--------------
Main Page
1) in the top Tabs, let's move it to the left, just after Config Page.
It could actually go first, since Config Page shows up at startup and then
is not used for an entire night.

2) Filter and Grating Status panel
Hard to say how this work in simulated mode. The "Set Filter" and "Set Grating" buttons are grayed out. If selecting a filter in the dropdown menu is enough to move the wheel, then the Set buttons can be removed (not shown).

3) Science panel
I would issue a warning when the Nr. of Frames is "large", or the exposure time is "long", and especially when  their product makes an exposure longer than say 600s. Errors are possible and one can wait up to some point, e.g. 600s, before shutting down and re-initiating the system.

4) Take Exposure Panel
Again, let's remove the "Flip on Load" button.

5) Fits Manager Panel
I have added a directory ./x_test with test material for the /R136 cluster. In the /fits_original_unflipped_basicHDR with an
original image (cluster R136) as returned from SISI; the image is flipped in X axis and has the original bare header of SISI. The Fits file, however, has been cleaned with the old "silly" routine.
I have added a second directory /fits_original_unflipped_QLHDR containing the same image with the Fits header modified by the current version of the GUI. It is clear that a number of parameters are missing, in particular those that we can grab from the telescope telemetry and our own WCS solution. Also, if there is an associated slit mask loaded in the DMD we should add the name. Having the mask pixel file would also make a nice extension.

6) To continue after having loaded the R136 image, I try to enter the coordinates
RA: 5.647025
DEC:-69.101911
but I get an error:entering the RA value fills also identically the DEC, and viceversa filling the DEC sets the RA.
=> FIXED

7) Then I move to the right and work with the RegionFiles I copied in the R136.
I load the RMC 136-T00_RADEC=84.6787022-69.1007396.reg file, but when I push
Get Centre/Point from Filename, I still see the RADEC of NGC3105.
=> FIXED

8) I had to check the twirl WCS code and managed to get it running. However, there is still work to do in what concerns the communication with SOAR: if we have SOAR connected, the nominal RADEC must be read from the telescope.

9) Once the images has been "twirled", we need to write the WCS solution into the fits file, creating a new version. Let's say this is the QL version. Also the other data from telescope telemetry can go in the fits header at this point.

10) We have learned at SOAR how the pointing is actually made, and it is instructive to repeat it here.
- During daytime calibration, we measure where the [0,0] position of the Guide Star probe is in our field. It will generally be not at the center of our field, but at a pixel (x0,y0).
- When we point the telescope to a target of known coordinates (RA0,DEC0), we want to put our target at the points (x0,y0). In this way, we know that point
(x0,y0) has coordinates (RA0,DEC0).
- The operator will find a good guid star at (RAg,DECg), and will therefore
convert the difference (RAg-RA) and (DECg-DEC0) in an offset move for the GS probe, that will nicely find the guide star.
- Where is the problem? When the telescope points our target coordinates (RA0,DEC0), it will not get exactly there. The (x0,y0) pixels will look at another point on the sky, a couple of arcseconds away. We need to know the RADEC of that point to apply an offset to the telescope and have the match between (x0,y0) and our commanded position.
- Therefore we need to do a couple of things:
A) get the values of pixel (x0, y0). They can be entered in the form at fields GS(0,0)
B) get the WCS of our image, and see where the (RA0,DEC0) coordinates fall; they will be at pixel (x1, y1).
C) Move the telescope so that RA0,DEC0 fall on (x0,y0).
D) Notes:
- the move is done in arc seconds, not in pixels. It is important to know the scale
- the image is shown reverted, but the WCS solution is intelligent and can return the right offset to be applied. But we need to be sure about this because we lost hours (due to the fact that the image was flipped here and there in the code, bad idea).
E) Changes to the GUI:
- add the fields to enter/display the pixels (x0, y0) of the GS home. These values are supposed to be constant as long as the instrument is not moved from the telescope. Can be placed in the GS page.
- add the fields to display where the nominal coordinates (RA0,DEC0) fall in the field after WCS solution. These are also pixels (x1,y1).
- The CNTR RA: and CNTR DEC: fields can be removed. We don't really care about the center of the field, what is key is the pixel (x0,y0).
- We can display the difference (x0-x1, yo-y1) in arc seconds. They can substitute the X Shift (mm) and Y Shift (mm) offset in the GSPage. Again, the move is done in arc seconds.
- Finally, add a button "Center Telescope" that sends to the telescope the offset signal, so (RA0,DEC0) falls exactly on (x0,y0)